Moved to `netutil` name. Moved from `os` to `env` pkg name.
