+++ integration_tests/pkg/workloads/memcached_test.go
@@ -6,7 +6,7 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/intelsdi-x/swan/pkg/executor"
-	"github.com/intelsdi-x/swan/pkg/utils/os"
+	"github.com/intelsdi-x/swan/pkg/utils/env"
 	"github.com/intelsdi-x/swan/pkg/workloads/memcached"
 	. "github.com/smartystreets/goconvey/convey"
 )
@@ -25,7 +25,7 @@ func TestMemcachedWithExecutor(t *testing.T) {
 		config := memcached.DefaultMemcachedConfig()
 		// Prefer to run memcached locally using the current user,
 		// if it can be determined from the environment.
-		config.User = os.GetEnvOrDefault("USER", config.User)
+		config.User = env.GetOrDefault("USER", config.User)
 		memcachedLauncher := memcached.New(l, config)
 
 		Convey("When memcached is launched", func() {
