+++ pkg/utils/netutil/netutil.go
@@ -1,4 +1,4 @@
-package http
+package netutil
 
 import (
 	"net"
@@ -10,17 +10,7 @@ const retries = 5
 // IsListeningFunction is a function type for checking if http endpoint is responding.
 type IsListeningFunction func(address string, timeout time.Duration) bool
 
-// IsListeningMockedSuccess is a mocked IsListeningFunction returning always true.
-func IsListeningMockedSuccess(address string, timeout time.Duration) bool {
-	return true
-}
-
-// IsListeningMockedFailure is a mocked IsListeningFunction returning always false.
-func IsListeningMockedFailure(address string, timeout time.Duration) bool {
-	return false
-}
-
-// IsListening tries to connect to given address http address in a form of `http://ip:port`.
+// IsListening tries to establish TCP connection to given address in a form of `ip:port`.
 // It returns true when it was able to connect to given endpoint within timeout time.
 func IsListening(address string, timeout time.Duration) bool {
 	sleepTime := time.Duration(
