+++ pkg/workloads/memcached/memcached_test.go
@@ -5,11 +5,22 @@ import (
 	log "github.com/Sirupsen/logrus"
 	"github.com/intelsdi-x/swan/pkg/executor/mocks"
 	"github.com/intelsdi-x/swan/pkg/isolation"
-	"github.com/intelsdi-x/swan/pkg/utils/http"
 	. "github.com/smartystreets/goconvey/convey"
 	"syscall"
 	"testing"
+	"time"
 )
+
+// IsEndpointListeningMockedSuccess is a mocked IsListeningFunction returning always true.
+func IsEndpointListeningMockedSuccess(address string, timeout time.Duration) bool {
+	return true
+}
+
+// IsEndpointListeningMockedFailure is a mocked IsListeningFunction returning always false.
+func IsEndpointListeningMockedFailure(address string, timeout time.Duration) bool {
+
+	return false
+}
 
 // TestMemcachedWithMockedExecutor runs a Memcached launcher with the mocked executor to simulate
 // different cases like proper process execution and error case.
@@ -35,7 +46,7 @@ func TestMemcachedWithMockedExecutor(t *testing.T) {
 			memcachedLauncher := New(
 				mockedExecutor,
 				config)
-			memcachedLauncher.isMemcachedUp = http.IsListeningMockedSuccess
+			memcachedLauncher.isMemcachedUp = IsEndpointListeningMockedSuccess
 			Convey("While simulating proper execution", func() {
 				mockedExecutor.On("Execute", expectedCommand).Return(mockedTaskHandle, nil).Once()
 				mockedTaskHandle.On("Address").Return(expectedHost)
@@ -61,7 +72,7 @@ func TestMemcachedWithMockedExecutor(t *testing.T) {
 					Convey("When test connection to memcached fails task handle shall be nil and error shall be return", func() {
 						mockedTaskHandle.On("Stop").Return(nil)
 						mockedTaskHandle.On("Clean").Return(nil)
-						memcachedLauncher.isMemcachedUp = http.IsListeningMockedFailure
+						memcachedLauncher.isMemcachedUp = IsEndpointListeningMockedFailure
 						task, err := memcachedLauncher.Launch()
 						So(err, ShouldNotBeNil)
 						So(task, ShouldBeNil)
@@ -71,7 +82,7 @@ func TestMemcachedWithMockedExecutor(t *testing.T) {
 					Convey("When test connection to memcached fails and task.Stop fails task handle shall be nil and error shall be return", func() {
 						mockedTaskHandle.On("Stop").Return(errors.New("Test error code for stop"))
 						mockedTaskHandle.On("Clean").Return(nil)
-						memcachedLauncher.isMemcachedUp = http.IsListeningMockedFailure
+						memcachedLauncher.isMemcachedUp = IsEndpointListeningMockedFailure
 						task, err := memcachedLauncher.Launch()
 						So(err, ShouldNotBeNil)
 						So(task, ShouldBeNil)
@@ -81,7 +92,7 @@ func TestMemcachedWithMockedExecutor(t *testing.T) {
 					Convey("When test connection to memcached fails, task.Stop succeeds and task.Clean fails task handle shall be nil and error shall be return", func() {
 						mockedTaskHandle.On("Stop").Return(nil)
 						mockedTaskHandle.On("Clean").Return(errors.New("Test error code for clean"))
-						memcachedLauncher.isMemcachedUp = http.IsListeningMockedFailure
+						memcachedLauncher.isMemcachedUp = IsEndpointListeningMockedFailure
 						task, err := memcachedLauncher.Launch()
 						So(err, ShouldNotBeNil)
 						So(task, ShouldBeNil)
